Create tracepoint crate with low-level API
This PR splits the low-level tracepoint building blocks out into a
separate crate. The `tracepoint` crate exposes the `EventDataDescriptor`
and `TracepointState` types. These were previously part of the
`_internal` module of the `eventheader` crate.

- New `tracepoint` crate contains `EventDataDescriptor` and `TracepointState`.
- Update doc comments on these because they're now "public".
- eventheader crate now depends on `tracepoint` crate, no longer depends
  directly on `libc`, no longer has any native stuff.
- `write_eventheader` is no longer a method on `TracepointState`. It is
  now a free function (`TracepointState` moved to `tracepoint` crate,
  but `tracepoint` crate shouldn't do eventheader stuff).
- Fix clippy warnings reported in `eventheader_macros`.

Possible future work:

- Macro-based non-eventheader eventing system (`Provider` class and
  macros, similar to the `eventheader` crate).
- Runtime-dynamic non-eventheader eventing system (`Provider` and
  `EventBuilder` classes, similar to the `eventheader_dynamic` crate).
