+++ eventheader_dynamic/src/changelog.rs
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
 //! Release history
 
 #[allow(unused_imports)]
