+++ tracepoint/src/changelog.rs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+//! Release history
+
+#[allow(unused_imports)]
+use crate::*; // For docs
+
+/// # v0.4.0 (TBD)
+/// - Create tracepoint crate with low-level user_events API.
+pub mod v0_4_0 {}
